CinemaAPI/requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
elastic-transport==8.10.0
elasticsearch==8.10.1
fastapi==0.104.1
frozenlist==1.4.0
h11==0.14.0
idna==3.4
multidict==6.0.4
orjson==3.9.10
pydantic==2.4.2
pydantic_core==2.10.1
pydantic-settings==2.1.0
redis==5.0.1
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
urllib3==2.0.7
uvicorn==0.24.0.post1
uvloop==0.19.0
yarl==1.9.2
gunicorn==20.1.0
requests==2.31.0
opentelemetry-api==1.17.0
opentelemetry-sdk==1.17.0
opentelemetry-instrumentation-fastapi==0.38b0
opentelemetry-exporter-jaeger==1.17.0
python-jose==3.3.0
prometheus-fastapi-instrumentator